Overall update (mainly verifying Brownian motion)
The commit contains many updates and clean-up (Stokes.cu and Integrator.cu),
but the main change is related to the Brownian calculations.
Many errors are corrected and the Chow & Saad method is reimplemented.
Brownian motion is verified for a single sphere (Stokes-Einstein),
a pair of spheres (Boltzmann),
and a suspensions of spheres (Beenakker & Mazur 1984; Ladd 1990).

Credits to Boyuan Chen (Caltech) for help in the debugging.