Switch back to EU1.
There was a minor bug associated with the RK2 integrator under shear,
making the periodic images of the particles lagging by a distance
proportional to the product of their relative velocity and timestep.
In my tests, the difference between EU1 and RK2 for a sheared suspension
of shakers is very small, probably because the timestep is small
and particles interact with both real neighbors and their images.
This bug does not affect the dynamics when there is no shear (and depending
on how shear is imposed, boxResize or not, the effect may disappear).
In any case, RK2 is removed for potential errors.